rename tx to rgbpp_to_branch
